Implement more dom interfaces
git-svn-id: https://svn.lojjic.net/xul/trunk@539 8f8d9d7b-adf4-0310-9a85-ad1d9f542d00

+++ xul-dom-api/src/main/java/net/lojjic/xul/XULCommandEvent.java
@@ -16,7 +16,7 @@ public interface XULCommandEvent extends UIEvent {
 			boolean canBubbleArg,
 			boolean cancelableArg,
 			AbstractView viewArg,
-			long detailArg,
+			int detailArg,
 			boolean ctrlKeyArg,
 			boolean altKeyArg,
 			boolean shiftKeyArg,
